+++ src/templates/index.ts
@@ -1,3 +1,2 @@
-export * from "./ITemplate";
-export * from "./ITemplateEngine";
 export * from "./TemplateEngine";
+export * from "./TemplateEngineCollection";