+++ src/tree/items/SolutionFileTreeItem.ts
@@ -2,8 +2,8 @@ import { ProjectInSolution } from "@core/Solutions";
 import { TreeItem, TreeItemCollapsibleState, TreeItemContext, ContextValues } from "@tree";
 
 export class SolutionFileTreeItem extends TreeItem {
-    constructor(context: TreeItemContext, name: string, filepath: string, public readonly projectInSolution?: ProjectInSolution) {
-        super(context, name, TreeItemCollapsibleState.None, ContextValues.solutionFile, filepath);
+    constructor(context: TreeItemContext, name: string, filepath: string, projectInSolution: ProjectInSolution) {
+        super(context, name, TreeItemCollapsibleState.None, ContextValues.solutionFile, filepath, projectInSolution);
     }
 
     command = {